[!] чтение определений индексов из БД [-] ошибка чтения определений табличных типов из БД

+++ cmd/engine/sqlserver/constraints.go
@@ -32,7 +32,7 @@ type Index struct {
 	// IgnoreDupKey значение параметра IGNORE_DUP_KEY
 	IgnoreDupKey bool
 	// FillFactor значение FILLFACTOR
-	FillFactor int
+	FillFactor byte
 	// IsPadded значение параметра PADINDEX
 	IsPadded bool
 	// IsDisabled индекс отключен
